add some basic tests, restructure project
the compiler part is now built as a library, with a "driver" component, which is responsible for running a "compiler instance". main basically just forwards to that.
This allows linking against the compiler (and invoking individual APIs, like the mangler, from the test target w/out having to somehow work around the fact that we'd have 2 main functions)